Fix how namespace matching happens in SymfonyFileLocator
The problem was that namespace prefixes can overlap, and for classes
that would match more than one prefix only the first prefix would be
tested for. This meant that the findMappingFile and fileExists functions
would return different results depending on the order of the stored
prefixes.

For these methods to return the correct values the prefixes must be
listed in order of most to least specific. For example

    ['Vendor\Package\Entities', 'Vendor\Package\Entities\Blog']

would return the wrong result whereas

    ['Vendor\Package\Entities\Blog', 'Vendor\Package\Entities']

would return the correct one.