Reach PHPStan level 7
We have found that there were lots of conflicts during the last merge up
to 3.0. Reaching level 7 on lower branches reduces differences between
them and thus the amount of conflits for future merges. It should also
ensure that no new PHPStan error are discovered when merging up because
they are detected at level 7 and not at level 5.