Fix AbstractManagerRegistry::getManagerForClass when proxy class is used.
Fixes: https://github.com/doctrine/persistence/issues/293

This PR passes the correct class name to the Metadata Factory `isTransient` method whether the original class is a proxy class or a real class. Whenever a proxy class is the original class, `isTransient` will get the parent class name and whenever a real class is the original class, `isTransient` will get the real class name.