Throw an exception when passing an entity alias
Many users have missed the deprecation about using an entity alias when
they migrated to doctrine/persistence 3.
Passing an entity alias was triggering a weird error due to generating
an invalid PSR-6 cache key due to `:` being a reserved keyword. So this
adds a dedicated check reporting that it is an invalid class name
instead.