Feature: allow `ColocatedMappingDriver` to accept iterable source file path names
Before it was only possible to pass an array of paths where to look for. Now, it's possible to explicitly provide fine-grained iterable of only necessary files to the Driver. It's anticipated to use Symfony Finder in the Bundles using this library. This gives much flexibility for client code to configure which files should be included and which should not.