Refactor: simplify `RegexIterator` in `ColocatedMappingDriver`
`i` flag is removed - it makes search case-insensitive. There's no point in making search case-insensitive for an extension regex (`.php`), since the extension is always lowercase.

`GET_MATCH` is changed into `MATCH` - using `MATCH` is more obvious, since there are **NO implicit modifications** of the result as in the case of `GET_MATCH`, which returns only the "matched" part of the path string instead of the actual `SplFileInfo` item.

`^.+` part of regex is removed - it's no longer necessary, since it had only been required for `GET_MATCH` to return the full path. Now, with `MATCH`, regex only affects results filtering and does not affect the results themselves.