Require PHP 8.4
This is consistent with other Doctrine projects. Note that I had to add
a "covariant" keyword on the template definition of getReflectionClass()
return type. That's because since PHP 8.4, ReflectionClass is considered
invariant because of lazy objects. Lazy objects introduce new methods to
ReflectionClass where the template appears in parameter position, for
instance that is the case for isUninitializedLazyObject().
I do not understand why this wasn't an issue with
ReflectionClass::isInstance(), which was introduced way earlier than
that though.

The solution I picked is to use "call-site variance", effectively saying
that this is OK in this instance (but it might not be, I'm not sure).

The alternative to adding this keyword would be to drop generic
information entirely.

See https://github.com/phpstan/phpstan/issues/12459 for more details.